feat: use `indexPages` for `docusaurus-search-local` (#123)
Enables `indexPages` but seems like theres some tweaks on it. I have
found this issue:
https://github.com/easyops-cn/docusaurus-search-local/issues/220

Where a user from 2022 using the same library had similar problems to
us, the thing is that FAQ contents are not being listed as part of
search bar results.

I will merge this as part of config and to test in production, but we
are likely to follow up with package maintainers and get back to this.