feat: support dollar sign cleaning on `copy=full` (#210)
Provides support for `FullText` removal of `$` and `>>` when found
maintaining format.

Unit tests are also provided and a cleanup in the structure so we have a
place for our own
scripts and custom logic.

The directory `functions` will be used to store our custom logic so we
can easily find it and import it into tests. Tests are executed now as
part of our CI process, to run tests
locally one can run `npm run test`.